cleaning folder of unnecessary files, rename folders
