+++ Frontend/src/app/app.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, async } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AppComponent } from './app.component';
+import { AppComponent } from './components/parityOfTheWeek/parityOfTheWeek.component';
 
 describe('AppComponent', () => {
   beforeEach(async(() => {
