+++ Frontend/src/app/components/parityOfTheWeek/parityOfTheWeek.component.css
